'cannot find stdout' fixed -- unzipping new agent
