--- main/Godeps/Godeps.json
@@ -1,82 +1,0 @@
-{
-	"ImportPath": "github.com/Azure/Guest-Configuration-Extension/main",
-	"GoVersion": "go1.10",
-	"GodepVersion": "v80",
-	"Deps": [
-		{
-			"ImportPath": "github.com/Azure/Guest-Configuration-Extension/pkg/seqnum",
-			"Comment": "v0.3.0-8-g9045c55",
-			"Rev": "9045c558d0eabef49520e0b319ce5b897c166ada"
-		},
-		{
-			"ImportPath": "github.com/Azure/azure-docker-extension/pkg/vmextension",
-			"Comment": "1.2.2-4-gd1d2547",
-			"Rev": "d1d2547150489b2495a8d09a642e7f515bbf484f"
-		},
-		{
-			"ImportPath": "github.com/Azure/azure-docker-extension/pkg/vmextension/status",
-			"Comment": "1.2.2-4-gd1d2547",
-			"Rev": "d1d2547150489b2495a8d09a642e7f515bbf484f"
-		},
-		{
-			"ImportPath": "github.com/davecgh/go-spew/spew",
-			"Comment": "v1.1.0-12-g8991bc2",
-			"Rev": "8991bc29aa16c548c550c7ff78260e27b9ab7c73"
-		},
-		{
-			"ImportPath": "github.com/go-kit/kit/log",
-			"Comment": "v0.7.0-25-ge2d71a0",
-			"Rev": "e2d71a06a40aa95cb82ccd72e854893612c02db7"
-		},
-		{
-			"ImportPath": "github.com/go-logfmt/logfmt",
-			"Comment": "v0.3.0",
-			"Rev": "390ab7935ee28ec6b286364bba9b4dd6410cb3d5"
-		},
-		{
-			"ImportPath": "github.com/go-stack/stack",
-			"Comment": "v1.7.0",
-			"Rev": "259ab82a6cad3992b4e21ff5cac294ccb06474bc"
-		},
-		{
-			"ImportPath": "github.com/kr/logfmt",
-			"Rev": "b84e30acd515aadc4b783ad4ff83aff3299bdfe0"
-		},
-		{
-			"ImportPath": "github.com/mcuadros/go-version",
-			"Rev": "6d5863ca60fa6fe914b5fd43ed8533d7567c5b0b"
-		},
-		{
-			"ImportPath": "github.com/pkg/errors",
-			"Comment": "v0.8.0-12-g816c908",
-			"Rev": "816c9085562cd7ee03e7f8188a1cfd942858cded"
-		},
-		{
-			"ImportPath": "github.com/pmezard/go-difflib/difflib",
-			"Comment": "v1.0.0",
-			"Rev": "792786c7400a136282c1664665ae0a8db921c6c2"
-		},
-		{
-			"ImportPath": "github.com/stretchr/testify/assert",
-			"Comment": "v1.2.2",
-			"Rev": "f35b8ab0b5a2cef36673838d662e249dd9c94686"
-		},
-		{
-			"ImportPath": "github.com/stretchr/testify/require",
-			"Comment": "v1.2.2",
-			"Rev": "f35b8ab0b5a2cef36673838d662e249dd9c94686"
-		},
-		{
-			"ImportPath": "github.com/xeipuuv/gojsonpointer",
-			"Rev": "4e3ac2762d5f479393488629ee9370b50873b3a6"
-		},
-		{
-			"ImportPath": "github.com/xeipuuv/gojsonreference",
-			"Rev": "bd5ef7bd5415a7ac448318e64f11a24cd21e594b"
-		},
-		{
-			"ImportPath": "github.com/xeipuuv/gojsonschema",
-			"Rev": "1d523034197ff1f222f6429836dd36a2457a1874"
-		}
-	]
-}
